Default to using i32 everywhere until needed
I know it's not always the most natural type given the values
represented, but it's a sane default and I'm not too worried about
optimization at this point.

Keeping things consistent helps too, considering we're doing a bunch of
math with these values, there are proper assertions in place, and then
we don't need to constantly cast variables.