revert: use convention of a const Self within container-level scope
I didn't like this. While it does reduce repetition of the type name, it
makes it harder to orient yourself when reading a function in isolation
without having to scroll around and see what struct you're inside of.

I'm also toying with using `rg 'pub fn' src/` as a quick way to generate
a bare-bones "repository map" [1], and having the explicit types in the
function signatures would help to provide useful context.

[1] https://aider.chat/docs/repomap.html