feat: support correct whole-tone intervals
Much like the chromatic scale, the intervals are different based on
the tonic, so we use a lookup table. I tried to generalize this so the
lookup table wouldn't be necessary, but it was much more complex and
hard to follow.

I've also introduced a mode field on the scale, but it's not being used
properly just yet.