feat!: implement modes as scale patterns
I misunderstood what modes were doing, and although the notes appear
to be rotating, it depends on the tonic and each mode has it's own
dedicated pattern. That being the case, we don't need a mode field
on the struct itself, we can just have named patterns for the various
modes.