fix!: follow upstream rename of CompTimeStringMap
Not only did the name change, but the initialization procedure and
types also needed to be tweaked.

This made me realize that with my other memory management changes, the
attempts to lookup shorthands for chromatic and whole-tone patterns was
failing. That was due to the Pitch.asText() return value going out of
scope, so it was always blank. Instead of trying to use a static array
and be tricky, we'll pass in an allocator like I probably should have to
begin with.

See:
- https://github.com/ziglang/zig/pull/19719