docs: update screenshot with latest layout
I just realized that raylib-zig does not have support for raygui, so my
plan of using those features isn't going to work. I may try a different
framework, but I'll have to tinker with this quite a bit either way. I'm
also not a huge fan of the font situation with raylib, so I wanted to
capture the current state of things.